writing: move shared io helper to plugin scripts dir, drop word count
